deploy/requirements.txt
PyYAML>=6.0
libvirt-python>=9.0.0

